Gro-Hub/requirements.txt
flask
gunicorn
joblib
pandas
scikit-learn
numpy
pdfplumber
python-docx
